Add regex_match to hook functions
Summary:
Add a helper function for hooks that can be used to match a string to a regex.

Regex matching can either be called from ctx.regex_match(s, r) or file.path_regex_match(r)

This will later be used in the no_bad_filenames hook, but will be helpful for other
hooks too.

Reviewed By: purplefox

Differential Revision: D12941544

fbshipit-source-id: f8f73a767da5ae77f486c925b6adc1333a198c3f