mononoke: remove lazy skiplist building
Summary:
Previously if a commit didn't have a skiplist index then it'd be built lazily.
So if a getbundle requests a very old commit then skiplist index entries for
too many commits might be built and stored in memory.

Instead of lazily building the index let's add a constructor that accepts a
skiplist graph. That'd allow us to skiplist graph persistently and create it on
demand

Reviewed By: jsgf

Differential Revision: D13169019

fbshipit-source-id: 813bdd7e6cc90312d97275a7db5e266d97b545e0