hgsubversion: add truncatemeta command
Summary:
In a future diff we want pushrebase to be able to rewrite commits
pulled from svn. In that situation, we need to be able to rewrite the svn
metadata so that the post-pushrebase commits are mapped to the svn rev numbers.

Let's add a truncatemeta command that will walk the latest commits in the
changelog and redo any that aren't mapped correctly.

Reviewed By: markbt

Differential Revision: D13074960

fbshipit-source-id: fe98879dd16cc0806d20ef6eab5f9d77f0e0c877