Add integration test for mononokeapi
Summary:
Add an integration test that talks to the API server from remotefilelog.

Since this involves running a local instance of Mononoke and the API server, I've added it to Mononoke's integration tests rather than Mercurial's integration tests in order to re-use the setup code from Mononoke's tests. Since the server is running locally, and would use a different SSL setup that the VIP that hg would normally access it through, the local API server is run without SSL enabled.

Reviewed By: Anastasiya-Zhyrkevich, farnz

Differential Revision: D13089196

fbshipit-source-id: 01f415d8ee7173f7f2ab3a234565fd79d618126e