mononoke: use RepoConfig for all cmd libraries
Summary:
Previously we manually specified blobstore type and all the necessary
parameters. That was error-prone but worked because we had only one blobstore.
Since we are going to add secondary blobstores soon configuring binaries like
blobimport will be harder because we'll need to specify parameters for all
blobstores. Let's make it so that
blobimport, mononoke admin and other binaries read the configuration the same
way as Mononoke does it i.e. via toml files.

Reviewed By: lukaspiatkowski

Differential Revision: D13183244

fbshipit-source-id: 99caa6348133acec11dd04ae44e1f9f0a8ebb197