remotefilelog: catch MissingNodesError exception in gc
Summary:
When garbage collecting lots of repository, a single corrupted
repository will stop the garbage collection for all the other repositories. In
the worst case, the first repository examined was bad, and thus no collection
will be done for the other repositories.

Let's log the faulty repository and continue with the other repositories.

Reviewed By: DurhamG

Differential Revision: D13316936

fbshipit-source-id: 9760710ff64d0173ae1c369991f438bbace11cf4