hgsubversion: use a single transaction for hgsubversion pull
Summary:
hgsubversion was doing one transaction per commit, which is both slow
and also causes a lot of packs to be created when operating on a remotefilelog
client.

Let's make it use a single transaction instead. Unfortunately the svn metadata
is not integrated with Mercurial transactions. If we're using a normal flat-text
revmap, if there is an exception when running pull, it will need to svn rebuild
the metadata to remove the bad data and continue.

If we're using a sqlite revmap, then we've integrated with the sqlite
transaction, so it should work as expected.

Reviewed By: phillco

Differential Revision: D13347708

fbshipit-source-id: 502c7e020ba2a2df1ec245a845b6d8eb8b9c3456