remotefilelog: fix rename traversals across multiple stores
Summary:
This bug has been here for 2+ years. Basically, when gathering the
ancestors for a given file node, if it traversed a rename it would lose track of
the new name and instead look up the new hash by the old name, which would fail.

We didn't hit this often, because it only causes a problem when you have partial
history and have to go fetch more history. In most remotefilelog repos we
download all of history, so you always have everything and therefore never hit
this.

Reviewed By: kulshrax, singhsrb

Differential Revision: D13332459

fbshipit-source-id: 120bfe9ac618a4979e1685f24dc6462fc7415b1b