infinitepush: fix visibility issue during update
Summary:
I faced an issue where I had the following DAG:

```
A   B
@   o
|  /
| /
|/
o C
```

I did the following operations at that point:

```
hg hide -r A
hg up top
```

but the update took me back to A instead of taking me to B as I expected.
quark-zju debugged the issue and we found out that this was due to D6899232.
This commit fixes the issue by ensuring that we do not make a filtered repo
unfiltered before the update operation.

Reviewed By: DurhamG

Differential Revision: D13388494

fbshipit-source-id: 1678f97a4b3179720c30cf1366e9301a2c27a5ba