setup: add a step to download build dependencies
Summary:
The way to download build dependencies are not consistent and reinvented in a
couple of places. For example:

- Cython: error out if it's not avaialable
- Rust: (the fb part of) distutils_rust auto-installs it
- Re2: Makefile and build_nupkg.py are doing the same thing
- Various Windows dependencies: downloaded by the fb part of build_nupkg.py

This is an attempt to unify part of them to be a command of setup.py:
"fetch_build_deps".

The motivation was to get IPython bundled. For now I just add re2 to the asset
list. Upcoming diffs will add IPython. Cython might be a good fit, too.

Reviewed By: DurhamG

Differential Revision: D13352622

fbshipit-source-id: 151d299663eba9bb49c9577be0e224f9de8a9912