setup: build and install IPython.zip at mercurial/thirdparty
Summary:
IPython is pretty handy (ex. `%timeit`, `?` etc). It should probably be the
recommended way to explore the code base for new people. This is a step towards
making it available for all platforms.

It's also smaller when compiled directly. The zip file containing all `.pyc`
files is just 8MB. When installed using CentOS's system package, a lot of GUI
dependencies will be installed, which need hundreds of MBs.

Note the zip only contains pure Python modules. The only native dependency
seems to be `scandir`. But `scandir` has a pure Python fallback. So it still
runs, just slower.

Reviewed By: DurhamG

Differential Revision: D13352617

fbshipit-source-id: 2ecbef69438ffc478389c5bec67bec5f83f7edeb