remove the non-toml config files
Summary:
Support for reading non-toml configs was removed in D10020958.
This updates the code to stop deploying these files.

The files in `/etc/eden/config.d` are only read by the python CLI code,
not by the edenfs daemon.  Therefore it should be fine to delete these files
even if some users still happen to still be running an older build of edenfs.

Reviewed By: chadaustin

Differential Revision: D13437953

fbshipit-source-id: 9001a49a697f80fc7a25343e3c22dff322924de8