mutation: add mutation information to commit contexts
Summary:
This is the first commit in a series that will replace obsolescence markers
with separate tracking of commit mutation and visibility.

Mutation information is metadata about which commit was mutated and in which
way that led to the creation of this commit.  It is embedded in the extras of
the commit object.

This commit adds the fields to the commit context.  Future commits will create
the mutation information when commits are mutated.

Reviewed By: DurhamG, quark-zju, ikostia

Differential Revision: D9975470

fbshipit-source-id: 0107308731dc021cb72734db8b5dc0a02e6e86e1