mutation: add split support
Summary:
Add support for mutation information being added when a commit is split.

The top of the final stack is the commit that stores the mutation metadata.
The other commits in the stack are just normal new commits, although the final
commit does record their hashes in its metadata.

Determining when the last commit in the stack happens before it is committed is
a little hard, as the code previously relied on the dirstate ending up clean
to detect a finished split.  Instead we look at the patches that come out of
the filter function and see if they match the original patches that were sent
in.

Reviewed By: DurhamG, ikostia

Differential Revision: D9975469

fbshipit-source-id: acec485f9b561952f4ccdbaaf9491c9d48a70f58