reduce false positives from eden doctor's rogue process check
Summary:
Fix the `eden doctor` check that searches for rogue Eden processes to be more
conservative, and stop reporting processes if it can't confirm if they are
legitimate or not.

This check is inherently racy, and cannot do a good job of identifying rogue
processes if other edenfs processes are in the middle of starting or shutting
down.  In practice this happens frequently when running the Eden tests,
especially in stress test modes where many tests are run repeatedly in
parallel.

This changes the code to only report rogue processes when it actually
identifies multiple edenfs processes trying to use the same state directory.

Reviewed By: strager

Differential Revision: D13373058

fbshipit-source-id: 0f0e9fdb59db6b9721afc4171d77b27239c4de94