remotefilelog: don't continously wrap generatefiles
Summary:
The setup function for remotefilelogserver wraps `cg1packer.generatefiles`,
however it does this each time a repo is set up.  After many repo objects
have been instantiated, we can end up with hundreds of wrappers around the
method, which overflows the stack when it is called.

Furthermore, in `wrappackers` we have replaced `cg1packer` with our own
`shallowbundle.shallowcg1packer`, so here remotefilelog is actually wrapping
itself!  We can just move the code for the server case into the remotefilelog
version of `generatefiles`.

Reviewed By: DurhamG, quark-zju

Differential Revision: D13505002

fbshipit-source-id: a236f7e62e6d2e5186f135bfe79477ce3e09e374