infinitepush: prefix exclude patterns with path:
Summary:
When pulling an infinitepush bundle from the server to a client, it
uses remotefilelog's excludepattern functionality to force the server to serve
filelogs to the client. Unfortunately, it didn't specify what type of pattern,
and therefore it was treated as a regular expression. This meant that any path
that happened to be a regular expression would generally not match itself,
resulting in data not being sent to the client.

The fix is to just prefix these with 'path:' since we know they are all exact
paths.

Reviewed By: phillco, quark-zju

Differential Revision: D13521814

fbshipit-source-id: 8afc35b37d5858913b80ed0babfa0b5e401f0ab4