stop making the mount point directory read-only
Summary:
Change `eden clone` so that it no longer makes the underlying mount point
directory read-only.

The original intention of making the directory read-only was to prevent users
from accidentally removing the `README_EDEN.txt` file that tells them how to
recover if edenfs is not running.  However in practice this ends up confusing
users that are legitimately trying to remove checkouts.  (At the moment
`eden rm` can sometimes time out trying to unmount the checkout, in which case
it can leave this directory behind.

Reviewed By: strager

Differential Revision: D13515854

fbshipit-source-id: c2e86632c1ab8a92d54179abb58839c68a8b23c6