scmutil: revise cleanupnodes behavior when moving bookmarks backwards
Summary:
Previously when a commit does not have a replacment, it will be moved fair back
to a commit that is not being replaced. For example, when A::C is being rebased
to A'::C' and B disappers due to being empty, the old code would move BOOK-B to
Z, while the new code would move BOOK-B to A':

  C             C'
  |             |
  B BOOK-B  ->  |
  |             |
  A      ------ A' <- new BOOK-B
  |     /
  Z ----           <- old BOOK-B

Note, the current `rebase` implementation overrides the "moves" calcuation used
in cleanupnodes. It already has the new behavior. So there are no changes in
rebase tests.

Right now, the real intended user is absorb, so I'm not adding new tests here.
Without this change, when absorb migrates to cleanupnodes, its test will break.
So I'm not adding new tests here.

Reviewed By: DurhamG

Differential Revision: D13531964

fbshipit-source-id: 03b6afa116e1a7b08b33a2c8856f2e52d6f8043a