treemanifest: add test demonstrating problem with linknodes
Summary:
When we send a bundle to the server, the treemanifest nodes are sent with
whatever linknode they have.  This linknode might not refer to a commit that
the server knows about, which makes the bundle invalid.  Add a test that
demonstrates that the server aborts in this case.

Pushrebase works fine as it rewrites the commits, generating new linknodes for
all of the trees.

Reviewed By: quark-zju

Differential Revision: D13504565

fbshipit-source-id: 39894d367c111aea5cef7de2d7da122e39d9debe