treemanifest: fixup linknodes when sending to the server
Summary:
When sending a bundle to the server, ensure that the treemanifest linknodes
correctly refer to a commit that is in the bundle.

Treemanifest ignores the linknodes of the subtrees - instead they inherit the
linknode of the root manifest in which they were introduced, so we only need to
fix up the linknode for the root manifest and propagate that to the subtrees.

Reviewed By: quark-zju

Differential Revision: D13504564

fbshipit-source-id: 2f481a4939239784d84d5db12c70d473b3045610